Updates structure
