+++ pgmpy/datasets/__init__.py
@@ -1,9 +1,11 @@
-from ._registry import DATASETS, dataset_class
-from .datasets import BaseDataset, load_dataset
+from ._base import DATASETS, DatasetRegistry, _BaseDataset, dataset_class, load_dataset
+from .abalone import Abalone  # noqa: F401
+from .sachs import Sachs  # noqa: F401
 
 __all__ = [
+    "_BaseDataset",
     "DATASETS",
+    "DatasetRegistry",
     "dataset_class",
-    "BaseDataset",
     "load_dataset",
 ]
